TeacherCreated ok, refactor TeacherProfile

+++ src/main/java/com/education/rural/persistence/repository/Identity/specialization/teacher/profile/TeacherProfileEntityRepository.java
@@ -12,50 +12,49 @@ import java.util.List;
 @Repository
 public class TeacherProfileEntityRepository implements TeacherProfileRepository {
 
-private final JpaTeacherProfileRepository jpaTeacherProfileRepository;
-private final TeacherProfileMapper teacherProfileMapper;
+    private final TeacherProfileMapper teacherProfileMapper;
+    private final JpaTeacherProfileRepository jpaTeacherProfileRepository;
 
-    public TeacherProfileEntityRepository(JpaTeacherProfileRepository jpaTeacherRepository, TeacherProfileMapper teacherProfileMapper) {
-        this.jpaTeacherProfileRepository = jpaTeacherRepository;
+    public TeacherProfileEntityRepository(TeacherProfileMapper teacherProfileMapper, JpaTeacherProfileRepository jpaTeacherProfileRepository) {
         this.teacherProfileMapper = teacherProfileMapper;
+        this.jpaTeacherProfileRepository = jpaTeacherProfileRepository;
     }
-
 
     @Override
     public List<TeacherProfileDto> findAll() {
-        return this.teacherProfileMapper.toDtos(this.jpaTeacherProfileRepository.findAll());
+        return this.teacherProfileMapper.toDtos(jpaTeacherProfileRepository.findAll());
     }
 
     @Override
     public TeacherProfileDto findById(Integer id) {
-        return this.teacherProfileMapper.toDto(this.jpaTeacherProfileRepository.findById(id).orElseThrow(()->new RuntimeException("TeacherProfileId not found")));
+        return this.teacherProfileMapper.toDto(jpaTeacherProfileRepository
+                .findById(id).orElseThrow(()->new RuntimeException("TeacherProfile Not Found!")));
     }
 
     @Override
     public TeacherProfileDto findFirstByProfileNameIgnoreCase(String profileName) {
         return this.teacherProfileMapper.toDto(this.jpaTeacherProfileRepository.findFirstByProfileNameIgnoreCase(profileName));
+
     }
 
     @Override
     public TeacherProfileDto save(TeacherProfileDto teacherProfileDto) {
-        TeacherProfileEntity teacherProfileEntity = this.teacherProfileMapper.toEntity(teacherProfileDto);
-        TeacherProfileEntity savedProfile = this.jpaTeacherProfileRepository.save(teacherProfileEntity);
-        return this.teacherProfileMapper.toDto(savedProfile);
+        TeacherProfileEntity teacherProfileEntity = teacherProfileMapper.toEntity(teacherProfileDto);
+        TeacherProfileEntity savedTeacherProfileEntity = jpaTeacherProfileRepository.save(teacherProfileEntity);
+        return this.teacherProfileMapper.toDto(savedTeacherProfileEntity);
     }
 
     @Override
     public TeacherProfileDto patch(Integer id, UpdateTeacherProfileDto updateTeacherProfileDto) {
-        TeacherProfileEntity profileEntity=this.jpaTeacherProfileRepository.findById(id).orElseThrow(()->new RuntimeException("ProfileId not found"));
-        this.teacherProfileMapper.updateTeacherToDto(updateTeacherProfileDto,profileEntity);
-        return this.teacherProfileMapper.toDto(this.jpaTeacherProfileRepository.save(profileEntity));
+        TeacherProfileEntity teacherProfileEntity=this.jpaTeacherProfileRepository.findById(id).orElseThrow(()->new RuntimeException("TeacherProfile Not Found!"));
+        this.teacherProfileMapper.updateTeacherToDto(updateTeacherProfileDto,teacherProfileEntity);
+        return this.teacherProfileMapper.toDto(teacherProfileEntity);
     }
 
     @Override
     public void delete(Integer id) {
-        TeacherProfileEntity profileEntity=this.jpaTeacherProfileRepository.findById(id).orElseThrow(()->new RuntimeException("ProfileId not found"));
-        this.jpaTeacherProfileRepository.deleteById(id);
-        System.out.println("Teachers profile with id"+ id+" has been deleted ");
+        if (this.jpaTeacherProfileRepository.existsById(id)) {
+            this.jpaTeacherProfileRepository.deleteById(id);
+        }
     }
-
-
 }
